alphaforge/requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.10
mcp>=1.0.0
yfinance>=0.2.18
newsapi-python>=0.2.0
sec-edgar-downloader>=4.0.0
pytesseract>=0.3.10
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
chromadb>=0.4.0
neo4j>=5.0.0
redis>=4.5.0
fastapi>=0.100.0
uvicorn>=0.20.0
asyncio>=3.4.3
aiohttp>=3.8.0